// management/univention-portal/frontend/src/apis/notifications/api.ts
export interface Link {
    /**
     * The accepted URLs are limited to HTTP and HTTPS, they must be absolute URLs.
     * @type {string}
     * @memberof Link
     */
    'url': string;
    /**
     *
     * @type {string}
     * @memberof Link
     */
    'text'?: string;
    /**
     *
     * @type {string}
     * @memberof Link
     */
    'target'?: string;
}
// ...
export interface NotificationRead {
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'sourceUid': string;
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'targetUid': string;
    /**
     * The notification title shall be visible to the user in a highlighted way. The title should be kept short.
     * @type {string}
     * @memberof NotificationRead
     */
    'title': string;
    /**
     * The notification details shall be visible to the user below the title. The details can be longer than the title.
     * @type {string}
     * @memberof NotificationRead
     */
    'details': string;
    /**
     * The severity influences how the notification will be displayed to the user. Typically this influences the background color. It does not influence if the notification is shown or not.
     * @type {NotificationSeverity}
     * @memberof NotificationRead
     */
    'severity': NotificationSeverity;
    /**
     * A point in time at which the notification is not relevant anymore. After this point in time the notification shall not be presented to the user anymore.
     * @type {string}
     * @memberof NotificationRead
     */
    'expireTime'?: string;
    /**
     *
     * @type {Link}
     * @memberof NotificationRead
     */
    'link'?: Link;
    /**
     *
     * @type {string}
     * @memberof NotificationRead
     */
    'id': string;
    /**
     *
     * @type {boolean}
     * @memberof NotificationRead
     */
    'popup'?: boolean;
}
/**
 * An enumeration.
 * @export
 * @enum {string}
 */

export const NotificationSeverity = {
    Info: 'info',
    Success: 'success',
    Warning: 'warning',
    Error: 'error'
} as const;

export type NotificationSeverity = typeof NotificationSeverity[keyof typeof NotificationSeverity];